design:마이페이지 위치 찾기 ui수정
